feature migrate profile to new activity done
