+++ .idea/misc.xml
@@ -21,6 +21,7 @@
         <entry key="app/src/main/res/layout/activity_sign_in.xml" value="0.3378623188405797" />
         <entry key="app/src/main/res/layout/activity_sign_up.xml" value="0.307518115942029" />
         <entry key="app/src/main/res/layout/activity_update_password.xml" value="0.3145833333333333" />
+        <entry key="app/src/main/res/layout/activity_user_profile.xml" value="0.3541666666666667" />
       </map>
     </option>
   </component>
